+++ LivroDrama.h
@@ -11,7 +11,7 @@ private:
     bool capaDura;
 
 public:
-    LivroDrama(string nome, float valor, int qntEstoque, bool capaDura);
+    LivroDrama(string nome, int valor, int qntEstoque, bool capaDura);
 
     void setCapaDura(bool capa);
     void toString();
